requirements.txt
python-dotenv==1.0.1
spotipy==2.25.0
tqdm==4.67.1
